Refactored perform method into smaller components
