+++ src/main/java/org/fcrepo/client/MoveBuilder.java
@@ -17,12 +17,10 @@
 package org.fcrepo.client;
 
 import static org.fcrepo.client.FedoraHeaderConstants.DESTINATION;
-import static org.slf4j.LoggerFactory.getLogger;
 
 import java.net.URI;
 
 import org.apache.http.client.methods.HttpRequestBase;
-import org.slf4j.Logger;
 
 /**
  * Builds a request to move a resource (and its subtree) to a new location
@@ -31,33 +29,27 @@ import org.slf4j.Logger;
  */
 public class MoveBuilder<T extends MoveBuilder<T>> extends RequestBuilder<MoveBuilder<T>> {
 
-    private static final Logger LOGGER = getLogger(MoveBuilder.class);
-
     protected URI destinationUrl;
-
-    protected HttpMethods method;
 
     protected MoveBuilder(URI sourceUrl, URI destinationUrl, FcrepoClient client) {
         super(sourceUrl, client);
         this.destinationUrl = destinationUrl;
-        method = HttpMethods.MOVE;
-    }
-
-    @Override
-    public FcrepoResponse perform() throws FcrepoOperationFailedException {
-        final HttpRequestBase request = method.createRequest(targetUri);
-
-        if (destinationUrl != null) {
-            request.addHeader(DESTINATION, destinationUrl.toString());
-        }
-
-        LOGGER.debug("Fcrepo {} request of {} to {}", method.name(), targetUri, destinationUrl);
-
-        return client.executeRequest(targetUri, request);
     }
 
     @Override
     protected MoveBuilder<T> self() {
         return this;
+    }
+
+    @Override
+    protected HttpRequestBase createRequest() {
+        return HttpMethods.MOVE.createRequest(targetUri);
+    }
+    
+    @Override
+    protected void populateRequest(final HttpRequestBase request) {
+        if (destinationUrl != null) {
+            request.addHeader(DESTINATION, destinationUrl.toString());
+        }
     }
 }
